Integration of current limiting ! Untested + Made LRTTalon into a Configurable + Added configuration options 'forwardLimit' and 'reverseLimit' for current limiting (default 0.3 and -0.3) + Changed type of speed to double in CurrentLimit
git-svn-id: http://svn.lynbrookrobotics.com/code15@255 9f8849d2-cf13-4b18-a094-84ffabbe6057

+++ Robot/LRT15/src/com/team846/frc2015/actuators/LRTSpeedController.java
@@ -59,7 +59,7 @@ public abstract class LRTSpeedController extends Actuator
 	public abstract double GetHardwareValue();
 	protected abstract void Update();
 	public abstract void ConfigNeutralMode(LRTSpeedController.NeutralMode mode);
-	double CurrentLimit(double dutyCycle, float speed, float forwardLimit, float reverseLimit)
+	double CurrentLimit(double dutyCycle, double speed, double forwardLimit, double reverseLimit)
 	{
 		if (speed < 0)
 		{
